refactor(entity): extract frontmatter helpers to entity_frontmatter.go
entity_synthesizer.go was north of the 400-LOC soft target — frontmatter
parsing/application is a self-contained concern and moving it out keeps
the core synthesis loop focused on queue semantics + LLM shellout.
